% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_fungal_abundance.R
\name{fung_abund}
\alias{fung_abund}
\title{Remove low abundance counts, rarefy sample counts, calculate abundances per taxon, and
tissue types in taxmap object}
\usage{
fung_abund(obj)
}
\arguments{
\item{obj}{taxmap object containing tax_data and class_data tibbles}
}
\value{
taxmap object containing tax_data, class_data, taxon_abundance, and tax_sample tibbles
}
\description{
Remove low abundance counts, rarefy sample counts, calculate abundances per taxon, and
tissue types in taxmap object
}
