% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_fungal_abundance_tree.R
\name{fung_abund_tree}
\alias{fung_abund_tree}
\title{Visualize fungal abundance from all samples in taxmap object}
\usage{
fung_abund_tree(obj_complete)
}
\arguments{
\item{obj_complete}{taxmap object containing tax_data, class_data, taxon_abundance, and tax_sample tibbles}
}
\description{
Visualize fungal abundance from all samples in taxmap object
}
