% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_process_otu_dataset.R
\name{parse_tax}
\alias{parse_tax}
\title{Fix column typos, remove NAs, and parse taxonomic information in OTU dataset}
\usage{
parse_tax(otu_data)
}
\arguments{
\item{otu_data}{OTU dataset containing abundances and taxonomic classifications}
}
\value{
taxmap object containing tibble with OTU counts and tibble with class data for each taxon
}
\description{
Fix column typos, remove NAs, and parse taxonomic information in OTU dataset
}
